requirements.txt
boto3==1.26.152
botocore==1.29.152
pandas==2.0.1
python-dotenv==1.0.0
Requests==2.31.0
